cleaned it up
